requirements.txt
discord
spotipy
python-dotenv
flask
requests
